reward_app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
